lab 04

+++ labs/lab04/exercise2/lab-4-2.py
@@ -3,7 +3,7 @@ if income <= 50000:
     toaltax = income * 0
 else:
     if income <= 100000:
-        totaltax = float(income - 50000) * 1 / 100
+        totaltax = (income - 50000) * 1 / 100
     else:
-        totaltax = float(50000 * 1 / 100) + float(income - 100000) * 2 / 100
+        totaltax = (50000 * 1 / 100) + (income - 100000) * 2 / 100
 print(totaltax)
